null

+++ Hell/Hell/Contracts/Entities/IItem.cs
@@ -1,9 +1,9 @@
 ﻿public interface IItem
 {
     string Name { get; }
-    int StrengthBonus { get; }
-    int AgilityBonus { get; }
-    int IntelligenceBonus { get; }
-    int HitPointsBonus { get; }
-    int DamageBonus { get; }
+    long StrengthBonus { get; }
+    long AgilityBonus { get; }
+    long IntelligenceBonus { get; }
+    long HitPointsBonus { get; }
+    long DamageBonus { get; }
 }
